Ej10 bucles

+++ src/main/java/com/mireyaserrano/tema03/bucles/Ejercicio10.java
@@ -6,12 +6,12 @@ public class Ejercicio10 {
     public static void main(String args[]){
         Scanner scanner = new Scanner(System.in);
         System.out.println("Introduce un número: ");
-        int num = Integer.parseInt(scanner.nextLine());
-        int total = 0;
+        float num = Float.parseFloat(scanner.nextLine());
+        float total = 0;
         do {
             total = total + num;
             System.out.println("Introduce otro número: ");
-            num = Integer.parseInt(scanner.nextLine());
+            num = Float.parseFloat(scanner.nextLine());
         }while (num != 0);
         System.out.println("El resultado de la suma de los números hasta el 0 es: " + total);
     }
